+++ packages/amplify-graphql-ops-generator/src/generator/index.ts
@@ -1,0 +1,7 @@
+
+import * as types from './types'
+
+export * from './types'
+import generate  from './generate'
+export { generateMutations, generateSubscriptions, generateQueries } from './generateQueries'
+export default generate
