+++ packages/amplify-graphql-ops-generator/src/generator/utils/getType.ts
@@ -1,0 +1,10 @@
+import { GraphQLList, GraphQLNonNull, GraphQLType } from "graphql";
+
+import { GQLConcreteType } from "../types";
+
+export default function getType(typeObj: GraphQLType): GQLConcreteType {
+  if (typeObj instanceof GraphQLList || typeObj instanceof GraphQLNonNull) {
+    return getType(typeObj.ofType);
+  }
+  return typeObj;
+}
