+++ packages/amplify-graphql-ops-generator/src/generator/utils/isRequired.ts
@@ -1,0 +1,4 @@
+import { GraphQLArgument, GraphQLNonNull } from "graphql";
+export function isRequired(arg: GraphQLArgument): boolean {
+  return arg.type instanceof GraphQLNonNull ? true : false;
+}
