+++ packages/amplify-graphql-ops-generator/tests/generator/utils/isRequired.test.ts
@@ -1,0 +1,29 @@
+import { isRequired } from "../../../src/generator/utils/isRequired"
+import {
+  GraphQLScalarType,
+  Kind,
+  GraphQLInt,
+  GraphQLObjectType,
+  GraphQLString,
+  GraphQLNonNull,
+  GraphQLList,
+  GraphQLInputObjectType
+} from "graphql";
+describe("isRequired", () => {
+  const testObj = new GraphQLInputObjectType({
+    name: "Address",
+    fields: {
+      street: { type: GraphQLString },
+      requiredInt: { type: new GraphQLNonNull(GraphQLInt) },
+    }
+  });
+
+  it("should return false for null types", () => {
+    expect(isRequired(testObj.getFields().street)).toEqual(false);
+  });
+
+  it("should return true for non null types", () => {
+    expect(isRequired(testObj.getFields().requiredInt)).toEqual(true);
+  });
+
+});
