+++ packages/amplify-graphql-ops-generator/tsconfig.json
@@ -17,6 +17,7 @@
     },
     "exclude": [
         "node_modules",
-        "lib"
+        "lib",
+        "tests"
     ]
 }
